filelist.sha1: write as-if --binary was passed to sha1sum tool.
The asterisk matters. It signifies that the file in question
should be read in binary mode. Which is rather true for these
raw files.

This did not seem to matter in practice so far,
and i kinda missed it the first time i looked at this, but
better do the right thing..