Make `/data/` generation atomic
Instead of deleting the old directory and then creating a new one,
and thus having a downtime with partially missing data,
instead create the new data in a timestamped directory,
and then just change the `/data/`'s symlink to point to said new dir.

This still may result in data going missing in the middle of user sync,
but only for the files that got unmarked as verified, but not deleted.
This basically never happens.
If file is deleted, it goes missing immediately.